Pagination And Sorting has been added

+++ src/main/java/com/example/keyrun/ECom/Repository/CategoryRepo.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface CategoryRepo extends JpaRepository<Category, Integer>
+public interface CategoryRepo extends JpaRepository<Category, Long>
 {
 
     Category findByname(@NotBlank @Size(min = 5,message = "Category name must be contain at least 5 character") @Size(max = 20,message = "Category name must not be contain more than 5 character") String name);
